Move the main code to _lspro.py so that all the functions will get precompiled. It will make easier to create python script based on available functions.
- all item art tags move to itemart and all item info move to itemifo dictionary.
- After adding new source, refresh source right  and change to not showing settings dialog.
